Create Q1.Rmd
Principal Component Analysis:
The data for this exercise comes from a paper by Cortez, et al.
(2009) where the authors were trying to relate various chemical properties of red and white wine to
perceived quality. For this question, we will analyze only the data for the chemical properties, not the
quality. Also the original paper looked at red and white wine, we will only use the data for the red.

Using only scatterplots and the sample correlation matrices, summarize what you
believe to be are the most interesting associations you observe among these characteristics.

Perform a principal component analysis of this data using your preferred function.
As part of this analysis, please be sure to complete the following tasks:
a) Report the eigenvalues for all 11 principal components.
b) Use any two methods seen in the lectures to choose the number of principal components to
retain. Show any calculation that you might need to use the methods.
c) For the first two principal components, plot and interpret components in terms of the original
variables. In particular, explain which variables are most highly correlated with each of these
two components and how these components are di